Moved to Github

+++ Makefile
@@ -3,6 +3,10 @@ PYTHON = python
 # these do not need changing
 FILE_WITH_VERSION = findimports.py
 FILE_WITH_CHANGELOG = CHANGES.txt
+VCS_STATUS = git status --porcelain
+VCS_EXPORT = git archive --format=tar --prefix=tmp/tree/ HEAD | tar -xf -
+VCS_TAG = git tag
+VCS_COMMIT_AND_PUSH = git commit -av -m "Post-release version bump" && git push && git push --tags
 
 .PHONY: default
 default:
@@ -37,13 +41,15 @@ dist:
 .PHONY: distcheck
 distcheck:
 	# Bit of a chicken-and-egg here, but if the tree is unclean, make
-	# distcheck will fail.  Thankfully bzr lets me uncommit.
-	@test -z "`bzr status 2>&1`" || { echo; echo "Your working tree is not clean" 1>&2; bzr status; exit 1; }
+	# distcheck will fail.
+ifndef FORCE
+	@test -z "`$(VCS_STATUS) 2>&1`" || { echo; echo "Your working tree is not clean" 1>&2; $(VCS_STATUS); exit 1; }
+endif
 	make dist
 	pkg_and_version=`$(PYTHON) setup.py --name`-`$(PYTHON) setup.py --version` && \
 	rm -rf tmp && \
 	mkdir tmp && \
-	bzr export tmp/tree && \
+	$(VCS_EXPORT) && \
 	cd tmp && \
 	tar xvzf ../dist/$$pkg_and_version.tar.gz && \
 	diff -ur $$pkg_and_version tree -x PKG-INFO -x setup.cfg -x '*.egg-info' && \
@@ -76,11 +82,11 @@ release: releasechecklist
 	# I'm chicken so I won't actually do these things yet
 	@echo "Please run"
 	@echo
-	@echo "  $(PYTHON) setup.py sdist register upload && bzr tag `$(PYTHON) setup.py --version`"
+	@echo "  $(PYTHON) setup.py sdist register upload && $(VCS_TAG) `$(PYTHON) setup.py --version`"
 	@echo
 	@echo "Please increment the version number in $(FILE_WITH_VERSION)"
 	@echo "and add a new empty entry at the top of the changelog in $(FILE_WITH_CHANGELOG), then"
 	@echo
-	@echo '  bzr ci -m "Post-release version bump" && bzr push'
+	@echo '  $(VCS_COMMIT_AND_PUSH)'
 	@echo
